Remove `make clean` mention from the README
This is a superfluous information that doesn't bring much value on its own. People can either figuring out themselves if needed or simply ignore it (as the workdir directory cleaned by `make clean` is ignored via `.gitignore`).

The `make clean` target is mostly an extra safeguard for the CI. It doesn't serve much purposes for users and might even lead to confusion for future stuff we will work on shortly.