Remove the "Known issues" section from the README
We should keep the wiki as **the** source for documentation.
Duplicating some information in the README doesn't serve much purpose, isn't as "discoverable" and represents additional busy work.

We should keep mentioning known pending issues about `systemd` support in the `first-setup` script until they are solved though, as this is something that this image aims to provide out of the box.