Stop using tags for releases and clean old releases from the package registry after 90 days
It is a bit odd to use git tags for every release, when the code isn't
changing, it is just newer packages. We can simply rely on GitLab's package registry [1].

Also adding a new "cleanup" stage to delete old releases after 90 days from the package registry,
which matches the number of days the releases are kept on the mirrors (3 releases are kept on the mirrors
and we are releasing monthly).

[1] https://docs.gitlab.com/ee/user/packages/generic_packages/

Inspired by arch-boxes@d04c8274