Test the reproducibility of the image itself and add a dedicated script / CI stage for reproducibility related tests
In addition of testing the reproducibility of the image userspace (which we implemented in https://gitlab.archlinux.org/archlinux/archlinux-wsl/-/merge_requests/70), I think it'd be nice to test for the reproducibility of the image itself too (I have hopes that we'll be able to report a 100% reproducibility on that front as well as some point).

Given that this is now two different tests related to reproducibility, I think it warrants a dedicated script, make target and CI stage (for maintenance and discoverability).

As defined in https://gitlab.archlinux.org/archlinux/archlinux-wsl/-/merge_requests/70, failed reproducibility tests should not be blocking (at least for now), hence the `allow_failure: true`.